format elementwise op
