--- cinn/hlir/op/contrib/clip.h
@@ -1,35 +1,0 @@
-// Copyright (c) 2022 CINN Authors. All Rights Reserved.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-#pragma once
-
-#include <string>
-#include <vector>
-
-#include "cinn/ir/ir.h"
-#include "cinn/ir/ir_base.h"
-#include "cinn/ir/tensor.h"
-
-namespace cinn {
-namespace hlir {
-namespace op {
-
-std::vector<ir::Tensor> Clip(const ir::Tensor& in_tensor,
-                             const float max_val,
-                             const float min_val,
-                             const std::string& output_name);
-
-}  // namespace op
-}  // namespace hlir
-}  // namespace cinn
