+++ cinn/hlir/op/use_ops.h
@@ -23,16 +23,13 @@ CINN_USE_REGISTER(broadcast_grad_ops)
 CINN_USE_REGISTER(elementwise_ops)
 CINN_USE_REGISTER(transform_ops)
 CINN_USE_REGISTER(scatter_ops)
-CINN_USE_REGISTER(cast_ops)
 CINN_USE_REGISTER(sort_ops)
 CINN_USE_REGISTER(argmin_ops)
 CINN_USE_REGISTER(argmax_ops)
 CINN_USE_REGISTER(reduce_ops)
-CINN_USE_REGISTER(clip_ops)
 CINN_USE_REGISTER(custom_call_op)
 CINN_USE_REGISTER(arange_ops)
 CINN_USE_REGISTER(flip_ops)
-CINN_USE_REGISTER(gelu_ops)
 CINN_USE_REGISTER(repeat_ops)
 CINN_USE_REGISTER(one_hot_ops)
 CINN_USE_REGISTER(lookup_table_ops)
